+++ components/datepicker/datepicker.test.js
@@ -1,6 +1,6 @@
 import { createLocalVue, mount } from '@vue/test-utils';
 import { formatMonth } from '@/components/datepicker/utils.js';
-import DtDatepicker from './DtDatepicker.vue';
+import DtDatepicker from './datepicker.vue';
 import { MONTH_FORMAT } from '@/components/datepicker/datepicker_constants.js';
 
 const MOCK_DAY = 21;
